Falta mejorar el de búsqueda por fecha de las transacciones de inventario

+++ tienda/src/main/java/gt/edu/tienda/modelo/TransaccionInventarioDetalle.java
@@ -14,7 +14,7 @@ public class TransaccionInventarioDetalle {
 	
 	@Id
 	@Column(name = "idtransacciondet")
-	private int idTransaccionDet;
+	private Long idTransaccionDet;
 	
 	@Column(name = "idtransaccion")
 	private int idTransaccion;
