+++ tienda/src/main/java/gt/edu/tienda/repositorio/ITransaccionInventarioDetalleRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import gt.edu.tienda.modelo.TransaccionInventarioDetalle;
 
-public interface ITransaccionInventarioDetalleRepositorio extends CrudRepository<TransaccionInventarioDetalle, Integer> {
+public interface ITransaccionInventarioDetalleRepositorio extends CrudRepository<TransaccionInventarioDetalle, Long> {
 
 }
